+++ packages/create/template/blog/.vuepress/navbar/zh.ts
@@ -47,6 +47,6 @@ export const zhNavbar = navbar([
   {
     text: "V2 文档",
     icon: "note",
-    link: "https://theme-hope.vuejs.vuepress/zh/",
+    link: "https://theme-hope.vuejs.press/zh/",
   },
 ]);