+++ packages/theme/src/node/compact/helper-v2.ts
@@ -1,5 +1,4 @@
 import { type UserConfig } from "@vuepress/cli";
-import { type ThemeFunction } from "@vuepress/core";
 import { isFunction, isPlainObject } from "@vuepress/shared";
 import { colors } from "@vuepress/utils";
 
@@ -14,14 +13,6 @@ import {
 } from "../../shared/index.js";
 import { hopeTheme } from "../theme.js";
 import { logger } from "../utils.js";
-
-/**
- * import and use `hopeTheme` instead
- *
- * @description This function will be removed in v2 future release and is only available before v2 touch stable
- */
-export const hopeThemeLegacy = (themeOptions: ThemeOptions): ThemeFunction =>
-  hopeTheme(themeOptions, true);
 
 /**
  * @deprecated use `import { navbar } from "vuepress-theme-hope";` instead
@@ -120,10 +111,10 @@ export default {
 
   // check themeConfig
   if ("themeConfig" in config && isPlainObject(config["themeConfig"]))
-    config.theme = hopeThemeLegacy(config["themeConfig"] as ThemeOptions);
+    config.theme = hopeTheme(config["themeConfig"] as ThemeOptions);
 
   // check theme
-  if (!isFunction(config.theme)) config.theme = hopeThemeLegacy({});
+  if (!isFunction(config.theme)) config.theme = hopeTheme({});
 
   return config;
 };