+++ packages/theme/src/shared/options/layout/sidebar.ts
@@ -58,12 +58,12 @@ export interface SidebarLocaleOptions {
    * sidebar config
    *
    * @description Set to `false` to disable sidebar in current locale
-   * @see https://theme-hope.vuejs.vuepress/guide/layout/sidebar.html
+   * @see https://theme-hope.vuejs.press/guide/layout/sidebar.html
    *
    * 侧边栏配置
    *
    * @description 设置为 `false` 以在当前语言中禁用侧边栏
-   * @see https://theme-hope.vuejs.vuepress/zh/guide/layout/sidebar.html
+   * @see https://theme-hope.vuejs.press/zh/guide/layout/sidebar.html
    */
   sidebar?: SidebarOptions;